ct_dashboard_api: extract invitation sending logic to api layer
